Fix infinite recursion in delete-file-or-directory-internal
This function is supposed to be called on files that should not or can
not be trashed, and hence it is supposed to bypass the trash can and
unconditionally delete the file or directory. Recent changes to Emacs
caused it to use Emacs's trash functionality instead, which led to
this function being called again, recursively. Fix by let-binding
delete-by-moving-to-trash to nil.